pam: make (only) pam_unix_session do lastlog
illumos-gate bug #6057. This commit moves the responsibility for both
updating and reading the lastlog file to the pam_unix_session module,
and removes lastlog.h so that newly compiled third party software won't
interact with the file. This wasn't accepted in gate because of the need
to support old binaries. Since we don't have such a strong need we can
use this implementation and even the old filename; although that will
mean that existing lastlog writers will corrupt the file, I ultimately
think it's better to find those writers and fix them.