aw: remove aw in favor of letting gcc do the hard work
Instead of having aw be an overcomplicated (yet not thorough enough) wrapper
around what amounts to `gcc -E | gas`, let gcc do both pre-processing and
assembly by invoking it with -c.

For i86pc, there is only one difference in the amd64 unix binary:

    --- unix-before	Sat Jul 30 07:41:56 2016
    +++ unix-after	Sat Jul 30 07:41:56 2016
    @@ -23,7 +23,7 @@
         vpanic+0x29:  74 0e              je     +0xe	<vpanic+0x39>
         vpanic+0x2b:  48 8d 25 1e 9c 41  leaq   +0x419c1e(%rip),%rsp	<panic_stack>
                       00 
    -    vpanic+0x32:  48 81 c4 00 20 00  addq   $0x2000,%rsp
    +    vpanic+0x32:  48 81 c4 00 40 00  addq   $0x4000,%rsp
                       00 
         vpanic+0x39:  48 81 ec f0 00 00  subq   $0xf0,%rsp
                       00 

This corresponds to this instruction in vpanic (in i86_subr.s):

    addq   $PANICSTKSIZE, %rsp

So, the definition of PANICSTKSIZE changed from 8kB to 16kB.  Looking at the
definition of PANICSTKSIZE it becomes clear what has happened:

    #ifdef _LP64
    #define	PANICSTKSIZE	16384
    #else
    #define	PANICSTKSIZE	8192
    #endif

We have been assembling i86_subr.s all this time without defining _LP64 for
amd64.  Amusingly enough, since the panic stack itself is defined in C, the
stack itself is always the right size:

    $ /usr/bin/nm /platform/i86*/kernel/unix|grep panic_stack
    [4350]	|0xfec435f4|0x00002000|OBJT |GLOB |0    |10     |panic_stack
    [6791]	|0xf505c9f4|0x00002000|OBJT |GLOB |0    |10     |panic_stack
    $ /usr/bin/nm /platform/i86*/kernel/amd64/unix|grep panic_stack
    [3791]	|0xfffffffffbc79cb0|0x0000000000004000|OBJT |GLOB |0    |12     |panic_stack
    [4062]	|0xfffffffffbcbc390|0x0000000000004000|OBJT |GLOB |0    |12     |panic_stack

On i86xpv, the story is similar.  Not only is vpanic affected the same way
as on i86pc, but the definition of TRAP_INSTR changed - from being always
the i386 definition, to being amd64 on amd64.