uts: always use a valid C token for 3rd probe argument
This is a simple cleanup performed using the following shell script to make
it easier for Coccinelle to clean up the probes later.

    find . -type f -exec \
        gsed -i -e '/TNF_PROBE_/{s/\/\* CSTYLED \*\//""/}' {} \;

Coccinelle is having problems matching an empty argument, so this commit
replaces the empty argument with an empty string.  There is no difference in
the generated code since the third argument (if present) is simply
concatenated with another string literal.