libpam: support for /etc/pam.d
This commit is based on a patch by Joshua M. Clulow <josh@sysmgr.org>.  It
adds support for /etc/pam.d and it splits up /etc/pam.conf into a number of
shards.  Unlike Joshua's work, this commit does not attempt to convert a
modified pam.conf file into shards.

The algorithm that decides which file to use when can be summarized with the
following pseudo-code:

  run_stack(service):
          if "/etc/pam.d/$service" exists:
                  load "/etc/pam.d/$service"
                  if stack is empty:
                          if "$service" != "other":
                                  include "/etc/pam.d/other"
                  else:
                          use loaded stack
          else:
                  load "/etc/pam.conf"
                  if stack is empty:
                          if "$service" != "other":
                                  include "/etc/pam.d/other"
                  else:
                          use loaded stack

illumos bug# 2947