env: remove dead/rarely used variables
The following variables are never used during a build and therefore can be
safely removed:

	ONBLD_BIN
	SUNW_NO_UPDATE_NOTIFY
	UT_NO_USAGE_TRACKING
	LINTDIRS

For no obvious reason, the env file states that setting RELEASE and
RELEASE_DATE is a possibility.  Nobody ever sets these via env files and so
these serve merely as clutter.  During release engineering, a custom env
file can be constructed to fix the release strings, or Makefile.master can
be updated on a release branch.  Either way, there is no reason to keep them
front and center for everyone to see.