vm: stop abusing p_hash for pvn tags
Instead of setting p_hash to a magic value to indicate that the page is
a pvn tag use a flag (P_PVN_TAG) in p_state.  While technically, we're
supposed to access p_state only while holding the proper locks, the
P_PVN_TAG is never added or removed from p_state (either it's there the
entire time the page_t exists or it's never there) and therefore we can
check it without any locking.