libdtrace: workaround missing tcph_t from ctf
Now that nca was removed in commit
4de7b7d44ed5ed5e9b1cf1dc3ad5c0cd393660e3, it seems that gcc omits DWARF
for the tcph_t type, and that breaks dtrace as an unfortunate side
effect since it relies on the existence of that type. To work around
this we could add -fno-eliminate-unused-debug-types to some kernel
compilation unit that happens to include inet/tcp.h, but that may pull
in more unused types. For now, let's just drop the typedef into the
dtrace consumer.