kernel: fix ioctl mis-interpreting file flags as data model
A long time ago, someone decided that the process data model information
should be passed in the upper 16 bits of the file flags passed into
fop_ioctl.  This blew up because of the recent expansion of ->f_flags to
include all the file flags, so we work around that for now.