mk: allow separate objdir for lib
for now this always sets MAKEOBJDIRPREFIX to /usr/obj/${MACHINE}. this
means that objects end up in /usr/obj/${MACHINE}/${.CURDIR} if that
directory exists.

the object directories will get used only if they exist and one of the
following is true:
 - make is recursing down from lib/ (lib/Makefile exports
   MAKEOBJDIRPREFIX to subdir makes)
 - an "obj" symlink exists in the directory make is called in. this may
   be created by 'make objlink' in subdirs, eg. lib/libfoo
otherwise, ${.CURDIR} will be used as objdir.

so, to build and install both 32 and 64-bit libraries on x86, you need
to run in lib/:

  make obj
  make MACHINE=amd64 obj
  make
  make MACHINE=amd64
  make install
  make MACHINE=amd64 install