nfs: nfs4 needs to fully initialize the nfs4_xattr_notsupp_vnode
The vnode initialization code is complicated because it is split between
several states due to the use of kmem cache constructors.  As a result,
any vnode that we call vn_reinit on must have already been initialized.

This fixes a regression caused by the recent merge of illumos-gate's
changes related to sdt DTrace probes in vnode hold macros.