tcp: remove _iss_incr tunable
It was introduced in illumos bug #917 in an attempt to address an issue
where high connection rates with the same connection-id tuple could
cause the new connection's ISN to not be greater than the old one
(breaking the connection reincarnation from TIME_WAIT heuristics).
However this does not apply to us since commit
915cb3b310e9b3a87ab2b352b22f57c23b34f41d; per RFC6528 we use a separate
ISN space for each (outgoing) connection-id tuple.

In fact the proposed algorithm in RFC6528 does not include the
per-connection component (ISS_INCR) at all, but mentions that
"Berkeley-derived kernels" used a per-second and a per-connection
component. Because we call gethrtime for each new outgoing connection,
the additional per-connection component should not be necessary.