boot: create_ramdisk should trust file(1)
The old Solaris file(1) used a built-in ELF parser to identify ELF files.
This meant that even without a magic file it knew how to process ELF files.
The new file(1) only uses magic files, and therefore specifying /dev/null as
its magic file makes it unable to identify ELF files as anything other than
"data".

create_ramdisk uses file(1) to determine if a file should be included on the
32-bit, 64-bit, or both ramdisks.