sys.mk: do not set -isysroot if DESTDIR is defined
DESTDIR is generally not something that is supposed to be set during
'make all', but we've used it up until this point. There is a use case
for avoiding to use headers and libraries from the currently running
system, relying only on things in objdirs or DESTDIR, but our bmake/mk
system does not currently support that anyway: -isysroot did that only
for headers (not libs) and only because those headers ended up DESTDIR
before 'bmake all' was invoked. We need a better solution for
new-objects-and-headers only builds; in the meantime -isysroot is a
problem for inter-component dependencies, so we fall back to the running
system for now.