Revert "ncurses: use SH instead of SHELL to hold path to /bin/sh"
This reverts commit 12bb93a511abdb55401565dd4432e8ca197af96f.

It turns out that the problem was that inside bldenv, SHELL was
/usr/bin/zsh *and* MAKEFLAGS was "e".  This lets the environment
override the assignment in the makefile.  (This isn't a concern for
nightly, since nightly uses 'env -i' to clean up the environment before
invoking bmake.)  Since as of 49c5cb15741a77034f1ccf7e493f73d9290870ae,
we always force bldenv to use /bin/sh, this is no longer a problem.