feature_tests.h: define _XPG7, _POSIX_C_SOURCE & __EXTENSIONS__ by default
this is a pretty large hammer that aims to make our default compilation
environment saner. there's a lot more work to do, but this will
hopefully make things easier.

__EXTENSIONS__ by default sounds a bit weird, but things break pretty
spectacularly if we only define XPG/POSIX without it - we have a bunch
of "interesting" checks like:
    #if !defined(_XPG4_2) || defined(__EXTENSIONS__)
which would probably be better hidden behind _UNLEASHED_VISIBLE or
something like that, if they're nonstandard symbols which we want
visible by default.