feature_tests.h: repurpose __UNLEASHED_VISIBLE
it is now set to 1 by default, 0 if user explicitly requests XPG or
POSIX (via _XOPEN_SOURCE or _POSIX_C_SOURCE), and it is always defined
after inclusion of feature_tests.h. this makes it more useful, a sort of
inverse replacement for __EXTENSIONS__ in illumos: these are nonstandard
interfaces that are available by default, but hidden if XPG or POSIX is
requested. illumos has kind-of-similar _STRICT_STDC/_STRICT_SYMBOLS, but
they do a subtly different thing and the existing checks in headers are
a mess between all three. so take a page out of the BSDs' book.