subdir.mk: build subdirs in parallel
this is a bit of a hack, and only applies if the target being built (eg.
all, cleandir) has not been defined prior to inclusion of subdir.mk (eg.
including prog.mk or lib.mk will keep the single-thread semantics). but
it does improve build speed, so use it as is at least until a more
comprehensive solution shows up.