libc: remove wordexp(3c)
it has been broken since ksh93 removal, and the implementation is
terrible: it spawns a shell for each invocation.

strictly speaking wordexp is specified by posix, but in practice it
doesn't seem to be required by most software; eg. openbsd doesn't have
it either (freebsd does, but they also use shell to implement it). let's
try the approach of just not providing this function.